requirements.txt
pandas
numpy
sqlalchemy
pymysql
apache-airflow